+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## Unreleased
+
+- Add `:menu` and `:menuitem` selector
+- Add `current:` filter for `:link` and `:link_or_button` selectors
+
 ## v0.8.2
 
 - Focus rich text when editing it